requirements.txt
pyTelegramBotAPI
google-generativeai
edge-tts
flask
requests
pillow